requirements.txt
# Core Data Analysis
pandas>=1.5.0
numpy>=1.23.0

# Statistical Analysis and Econometrics
statsmodels>=0.14.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.2.0

# Data Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Data Sources
pandas-datareader>=0.10.0
openassetpricing>=0.1.0

# Configuration
pyyaml>=6.0

# Excel Support
openpyxl>=3.0.0
xlrd>=2.0.0

# Jupyter Notebooks (for interactive analysis)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0

# Code Quality and Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Type Checking (optional)
mypy>=0.990

# Code Formatting (optional)
black>=23.0.0
flake8>=6.0.0
